+++ src/main/java/web/repository/ThemeRepository.java
@@ -9,7 +9,7 @@ public interface ThemeRepository {
 
     Long createTheme(Theme theme);
 
-    long deleteTheme(long themeId);
+    Integer deleteTheme(long themeId);
 
     Optional<List<Theme>> getThemes();
 }
